Ajustando arquitetura do serviço de usuario, com rotas de add Usuario proximo a ser concluida.
