+++ usuario-services/src/models/Item.js
